refactor: extract create link logirefactor: extract create link logic
